core/standards/html/semantics_structure_api/elements/content_models/kinds_of_content/interactive_content/link_click_close_window.html

<!doctype html>
<title>Link click closing window</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
var t = async_test();
window.name = "top_window";
onload = t.step_func(function() {
  var win = window.open("link_click_close_window-1.html");

  var w = new SharedWorker("link_click_close_window-1.js");
  w.port.postMessage("base");
  w.port.onmessage = t.step_func(
    function(e) {
      assert_equals(e.data, "PASS"); 
      t.done()
    })
});
</script>
